+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+* Remove the portfinder dependency and let ADB find an unused port instead
+
 ## 1.0.1 - 2018-09-01
 * Rename the `adbExecutable` configuration parameter